requirements.txt
pytest
pytest-playwright
pytest-bdd

